backend/requirements.txt
numpy==1.24.3
opencv-python==4.8.0
Pillow==10.0.0
scikit-image==0.21.0
scipy==1.11.2
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyWavelets==1.4.1